# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Book operations
  - name: chapters
    description: Chapter operations
  - name: generation
    description: AI curriculum generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookInput"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{bookId}:
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book's title, author, or grade
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookInput"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Book not found
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Book deleted
  /books/{bookId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Add a chapter to a book (with optional file upload)
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateChapterInput"
      responses:
        "201":
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /books/{bookId}/chapters/{chapterId}:
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter's title, page range, or number
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChapterInput"
      responses:
        "200":
          description: Chapter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
        "404":
          description: Chapter not found
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Chapter deleted
  /books/{bookId}/chapters/{chapterId}/workbook:
    get:
      operationId: getWorkbook
      tags: [generation]
      summary: Get the generated student workbook Markdown for a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workbook markdown content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
        "404":
          description: Not found or not yet generated
  /books/{bookId}/chapters/{chapterId}/teacher-guide:
    get:
      operationId: getTeacherGuide
      tags: [generation]
      summary: Get the generated teacher guide Markdown for a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher guide markdown content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
        "404":
          description: Not found or not yet generated
  /books/{bookId}/chapters/{chapterId}/regenerate:
    post:
      operationId: regenerateChapter
      tags: [generation]
      summary: Trigger regeneration of AI content for a chapter
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Regeneration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        grade:
          type: integer
        chapterCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - author
        - grade
        - chapterCount
        - createdAt
    CreateBookInput:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        grade:
          type: integer
      required:
        - title
        - author
        - grade
    UpdateBookInput:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        grade:
          type: integer
    Chapter:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        num:
          type: integer
          nullable: true
        title:
          type: string
        pages:
          type: string
        status:
          type: string
        date:
          type: string
          nullable: true
        file:
          type: string
          nullable: true
        hasWorkbook:
          type: boolean
        hasTeacherGuide:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - title
        - pages
        - status
        - hasWorkbook
        - hasTeacherGuide
        - createdAt
    CreateChapterInput:
      type: object
      properties:
        num:
          type: integer
        title:
          type: string
        pages:
          type: string
        status:
          type: string
        date:
          type: string
        file:
          type: string
      required:
        - title
        - pages
    UpdateChapterInput:
      type: object
      properties:
        num:
          type: integer
          nullable: true
        title:
          type: string
        pages:
          type: string
    GeneratedContent:
      type: object
      properties:
        markdown:
          type: string
        chapterId:
          type: integer
        type:
          type: string
      required:
        - markdown
        - chapterId
        - type